RecipeListView parent class - UNFINISHED
